完成vuecmf工具初的始化数据库、升级及回滚操作

+++ app/vuecmf/model/model_form_linkage.go
@@ -9,7 +9,7 @@ type ModelFormLinkage struct {
 	ModelFieldId uint `json:"model_field_id" form:"model_field_id" binding:"required" required_tips:"请选择" gorm:"column:model_field_id;size:32;uniqueIndex:unique_index;not null;default:0;comment:模型字段ID"`
 	LinkageFieldId uint `json:"linkage_field_id" form:"linkage_field_id" binding:"required" required_tips:"请选择" gorm:"column:linkage_field_id;size:32;uniqueIndex:unique_index;not null;default:0;comment:联动表单的字段ID"`
 	LinkageActionId uint `json:"linkage_action_id" form:"linkage_action_id" binding:"required" required_tips:"请选择" gorm:"column:linkage_action_id;size:32;not null;default:0;comment:获取联动表单数据的动作ID"`
-	Status uint `json:"status" form:"status"  gorm:"column:status;size:8;not null;default:10;comment:状态：10=开启，20=禁用"`
+	Status uint16 `json:"status" form:"status"  gorm:"column:status;size:8;not null;default:10;comment:状态：10=开启，20=禁用"`
 	
 }
